Refactor quarkdoc name transformers
