+++ quarkdown-quarkdoc/src/main/kotlin/com/quarkdown/quarkdoc/dokka/transformers/RenamingsStorer.kt
@@ -54,4 +54,11 @@ object RenamingsStorage {
      * @return the new name for the function with the given DRI, or null if it is not found.
      */
     operator fun get(dri: DRI): Renaming? = PrivateRenamingsStorage.renamings[dri]
+
+    /**
+     * Clears the stored renamings.
+     */
+    fun clear() {
+        PrivateRenamingsStorage.renamings.clear()
+    }
 }
